Fix audio. Added support for 2 audio streams.

+++ libv4l2cpp/inc/AnykaEncoderBase.h
@@ -36,7 +36,7 @@ public:
     void stop();
     bool encode();
 
-    size_t getEncodedFrameReadyFd() const;
+    int getEncodedFrameReadyFd() const;
     size_t getEncodedFrameSize() const;
     size_t getEncodedFrame(char* buffer, size_t bufferSize);
